Rename providerserver -> server + named loggers
