+++ internal/server/server.go
@@ -1,4 +1,4 @@
-package providerserver
+package server
 
 import (
 	"context"
@@ -12,15 +12,15 @@ import (
 )
 
 var (
-	_ pb.CSIDriverProviderServer = (*ProviderServer)(nil)
+	_ pb.CSIDriverProviderServer = (*Server)(nil)
 )
 
-// ProviderServer implements the secrets-store-csi-driver provider gRPC service interface.
-type ProviderServer struct {
+// Server implements the secrets-store-csi-driver provider gRPC service interface.
+type Server struct {
 	Logger hclog.Logger
 }
 
-func (p *ProviderServer) Version(context.Context, *pb.VersionRequest) (*pb.VersionResponse, error) {
+func (p *Server) Version(context.Context, *pb.VersionRequest) (*pb.VersionResponse, error) {
 	p.Logger.Info("Processing version method call")
 	return &pb.VersionResponse{
 		Version:        "v1alpha1",
@@ -29,7 +29,7 @@ func (p *ProviderServer) Version(context.Context, *pb.VersionRequest) (*pb.Versi
 	}, nil
 }
 
-func (p *ProviderServer) Mount(ctx context.Context, req *pb.MountRequest) (*pb.MountResponse, error) {
+func (p *Server) Mount(ctx context.Context, req *pb.MountRequest) (*pb.MountResponse, error) {
 	p.Logger.Info("Processing mount method call", "request", req)
 
 	versions, err := p.handleMountRequest(ctx, req.Attributes, req.TargetPath, req.Permission)
@@ -45,16 +45,16 @@ func (p *ProviderServer) Mount(ctx context.Context, req *pb.MountRequest) (*pb.M
 	return &pb.MountResponse{ObjectVersion: ov}, nil
 }
 
-func (p *ProviderServer) handleMountRequest(ctx context.Context, parametersStr, targetPath, permissionStr string) (map[string]string, error) {
+func (p *Server) handleMountRequest(ctx context.Context, parametersStr, targetPath, permissionStr string) (map[string]string, error) {
 	p.Logger.Debug("Handling mount request", "parametersStr", parametersStr)
 	cfg, err := config.Parse(parametersStr, targetPath, permissionStr)
 	if err != nil {
 		return nil, err
 	}
 
-	p.Logger.Debug("Running provider", "vault address", cfg.Parameters.VaultAddress, "roleName", cfg.Parameters.VaultRoleName)
+	p.Logger.Debug("Running provider server", "vault address", cfg.Parameters.VaultAddress, "roleName", cfg.Parameters.VaultRoleName)
 
-	provider := provider.NewProvider(p.Logger)
+	provider := provider.NewProvider(p.Logger.Named("provider"))
 	versions, err := provider.MountSecretsStoreObjectContent(ctx, cfg)
 	if err != nil {
 		return nil, fmt.Errorf("error making mount request: %w", err)
